+++ examples/telemetry_viewer.rs
@@ -1,13 +1,12 @@
 #!/usr/bin/env -S watchexec -r cargo run --example telemetry_viewer --features telemetry
 
-#[path = "../src/bin/gorbie-telemetry-viewer/app.rs"]
+#[path = "../src/bin/telemetry-viewer/app.rs"]
 mod app;
 
-use GORBIE::notebook;
 use GORBIE::NotebookCtx;
+use GORBIE::notebook;
 
 #[notebook]
 fn main(nb: &mut NotebookCtx) {
     app::notebook(nb);
 }
-